Add backend and channel changes
With the changes we have discussed a new system is needed to handle
the filtering requirements. We need for handling filtering based on
multiple criterias for multiple fields, and also to allow the same
filtering rule to apply to multiple backends, which will recieve
those messages specific to that critiera.