config: use ports-based config, deprecate ports in services
This is a minor redesign of the Honeytrap model, where the focus of the configuration moves from services (a list of services each with their
ports) to ports (a list of ports with their services, and a list of port-agnostic services).

As a consequence of this redesign, it is now possible to define a priority model for services on the same port. Simply put, services that come
first take a higher priority (i.e. are tested first for compatibility with the connection). An example application for this is running TR-069
and HTTP on the same port: TR-069 is HTTP-based but more specific, so it should be tested first, with HTTP being a fallback.